Added promoting role and updated methods in GarnetClusterController
In this commit, we have made several updates to the GarnetClusterController, including the addition of the `promoting` role and the implementation of a `RebalanceClusterAsync` method for rebalancing the cluster. We have also updated various methods to handle the `promoting` role and added a `RemoveLeavingAsync` method to handle nodes that are leaving.

In `garnetclusters.garnet.k8soperator.io.yaml`, the `promoting` role was added and the type of `slots` was changed from `string` to `integer`.

In `GarnetNode.cs`, methods were added to get the number of slots and to get the slots themselves.

In `GarnetRole.cs`, the `promoting` role was added.

In `GarnetClientExtensions.cs`, the `ForgetAsync` method was updated to handle the case where the node is not known.

In `GarnetHelper.cs`, methods were added to execute Redis commands and to get shards.

In `Test_ScalingOperations.cs`, tests were added to check the number of slots and to get the slots.

In `Shards.cs`, classes were added to represent shards and nodes, and a JSON converter was added for shards.

In `Test_Json.cs`, a test was added to deserialize shards.