Refactor GarnetClusterController and update models and tests
Significant refactoring has been done on `GarnetClusterController.cs` in the `GarnetOperator` namespace. The `GetShardsAsync` method has been separated from the `Reconciling` method and now retrieves shards and updates the cluster status. The `InitializeStatusAsync` and `ReconcileClusterAsync` methods have been updated to start new activities and calculate slots respectively. The `ConfigureClusterAsync` method now clones cluster nodes and calculates slots for each node.

The `Shards.cs` file in the `GarnetOperator.Models` namespace has been updated to replace the string type of the `Role` property in the `Node` class with the `GarnetRole` enum type. A new method `RoleFromString` has been added to convert a string role to a `GarnetRole` enum.

XML documentation comments have been added to the `ExecuteRedisCommandAsync` and `GetShardsAsync` methods in the `GarnetHelper.cs` file in the `GarnetOperator.Util` namespace.

Two new test methods: `CalculateSlots` and `GetSlotRanges` have been added to the `Test_ScalingOperations.cs` file in the `Test_GarnetOperator` namespace.

Two new classes `SlotMigration` and `SlotRange` have been added to the `GarnetOperator.Models` namespace. The `SlotMigration` class represents a migration of slots from one node to another, and the `SlotRange` class represents a range of slots. The `SlotMigration` class includes a `GetSlotRanges` method that calculates the ranges of slots to be migrated.