Refactor GarnetClusterController and update YAML files
Significant changes were made to the GarnetClusterController.cs file to improve efficiency and functionality. This includes the addition and removal of namespaces, modification of methods for better logging, exception handling, and node management, and the addition of health checks for pods. The return type of the GetClusterPodsAsync method was modified and a block of code updating the status of each pod was commented out.

Changes were also made to the garnetclusters.garnet.k8soperator.io.yaml and example.yaml files, including the reordering and removal of enum values, addition of a new property, and changes to the numberOfPrimaries and replicationFactor values.

Other updates include the removal of the GetPrimaryToReplicas method from the Cluster.cs file, addition of the System.ComponentModel namespace and update of the Role property in the GarnetNode.cs file, setting None as the default value for the GarnetRole enum in the GarnetRole.cs file, and updates to the GarnetClientExtensions.cs, GarnetHelper.cs, garnet-operator.csproj, and Test_Extensions.cs files.

A new file, ClusterNode.cs, was added in the GarnetOperator.Models namespace.