Updated solution items, constants, project reference, and using statements
Updated various files in the project. The `logo.png` file was added to the solution items in the `garnet-operator.sln` file. Constants in the `Constants.cs` file were updated with new values. An `Icon` attribute was added to the `Olm.cs` file, pointing to the `logo.png` file. The project reference in the `Test.garnet-operator.csproj` file was updated to point to the `src` directory. The `Test_Extensions.cs` file was updated to include additional `using` statements and the order of the `using` statements was rearranged.